server/requirements.txt
africastalking==1.2.8
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
coverage==7.6.1
cryptography==43.0.1
dj-database-url==2.2.0
Django==5.1.1
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-api-core==2.20.0
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
Markdown==3.7
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==24.1
proto-plus==1.24.0
protobuf==5.28.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
schema==0.7.7
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uuid==1.30
Werkzeug==3.0.4
whitenoise==6.7.0
